+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- [25-1-26] Add macros for MARK, SPOT, WIPE, RUN, SANDP.
 - [25-1-25-2] Add call history to REF CW and SSB.
 - [25-1-25-1] Fix:ratewindow failed to change stats when contest changed.
 - [25-1-25] Fix app not following radio mode. Fix crash on recalc mults.